fix: let docker assign ephemeral port and get from ContainerInspect
This was a tricky one; seems like the docker folks claim that the
ContainerInspect should be synchronous after container start.
See https://github.com/moby/moby/issues/42860#issuecomment-921772761

However, this isn't accurate, or at least I couldn't get it to
return the port detail on the first call to ContainerInspect after
the ContainerStart call. It seems to work consistently to assign
the port by second time it gets called, but for safety I added a
retry loop with a 1ms delay between each call to ContainerInspect.
I've never observed any retries beyond the first one.

This approach allows us to remove the getFreePort() hack, and we
I've removed the rety loop for the DialSSH() call, since we now
consistently get a working port mapping to allow the ssh dial to
connect.

Fixes #7.