refactor: split ClientConfig into ParseSSHConfig with methods
This replaces the old ClientConfig function which were multi-purpose
into a single ParseSSHConfig function that returns a sshConfig
object that can be queried to obtain the desired ClientConfig for
a host alias and ConnectAddr can be invoked separately to obtain
the connection string to be passed to the dial function.

This also improves the tests.