refactor: add support for StrictHostKeyChecking
This adds support for disabling host key checking, thus avoiding
the warning prompt when logging into a ssh server for which there
is no entry in the client's known_hosts file.

This option should only be disabled if you trust the hosting
infrastructure. See the shell script in issue #2 for an alternative
to create known_hosts file entries for a set of machines, which
will then be checked via the UserKnownHostsFile specified in the
ssh config file.