fix: CreateSSHGroup error handling and cleanup logic
This fixes the clean up logic for containers created by
CreateSSHGroup. We need to call ContainerRemove when we do our own
ContainerStop first (in this case AutoRemove: true does not work).
This also means that g.Close() must be called first (LIFO order)
to allow proper shutdown while the containers are still running
with network.