feat: extract logs from container on failure to aid diagnositics
These types of failurs are rare, and so this feature has not been
tested yet; nonetheless, it is there now so that it can help in
future debugging efforts.

The following code was tried as well, but didn't solve the problem:

func waitForSSH(addr string, timeout time.Duration) error {
	deadline := time.Now().Add(timeout)
	for time.Now().Before(deadline) {
		conn, err := net.DialTimeout("tcp", addr, 500*time.Millisecond)
		if err == nil {
			conn.Close()
			return nil
		}
		time.Sleep(200 * time.Millisecond)
	}
	return fmt.Errorf("timeout waiting for SSH at %s", addr)
}

The idea here was to wait for the container's ssh server to be live
before returning from createContainer(). It did not work, as the
rare problem show its ugly face on the first run after introducing
the above code.