refactor: use new ConsumeStream helper from container module
This makes use of the ConsumeStream function to parse the JSON responses
from the docker engine and print them to t.Output() instead of stdout
so that we can control whether to emit output using the -v flag when
running tests. ConsumeStream will parse the JSON and understands if
the build failed with an error and can print the output provided by
the docker engine to assist with debugging any docker build issues.